Airfield intake working

+++ src/com/skilldistillery/jets/entities/Jet.java
@@ -10,7 +10,7 @@ public abstract class Jet {
 	public Jet() {}
 	
 	
-	public Jet(String model, int speed, int range, double price) {
+	public Jet(String model, double speed, int range, long price) {
 		super();
 		this.model = model;
 		this.speed = speed;
